Bugfix when no type flavor text is available
Before the app would crash when there wasn't any flavor text available for the type of artwork. I saw this on /artwork/17de0063-940d-4a9b-8983-993da55db4dc where the type is "ceramics" but the entry is for "ceramic".
I thought we covered every type in the dataset. While we could fix the data where there are these one-off issues, it's probably better to have this error handling. Now the flavor text is only set when there is some available.